Switch to v2 of the google api
