+++ .gitignore
@@ -1,3 +1,5 @@
+.idea/
+
 # Binaries for programs and plugins
 *.exe
 *.exe~
