+++ fs.go
@@ -3,23 +3,35 @@
 package s3fs
 
 import (
+	"context"
 	"errors"
 	"io/fs"
 	"path"
 
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/awserr"
-	"github.com/aws/aws-sdk-go/service/s3"
-	"github.com/aws/aws-sdk-go/service/s3/s3iface"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/s3"
+	smithyhttp "github.com/aws/smithy-go/transport/http"
 )
 
 var (
 	_ fs.FS        = (*S3FS)(nil)
 	_ fs.StatFS    = (*S3FS)(nil)
 	_ fs.ReadDirFS = (*S3FS)(nil)
+	_ S3Api        = (*s3.Client)(nil)
 )
 
 var errNotDir = errors.New("not a dir")
+
+// S3Api is a api interface for all functions needed for this project.
+// This is suggested by https://aws.github.io/aws-sdk-go-v2/docs/unit-testing/#mocking-client-operations
+type S3Api interface {
+	HeadObject(ctx context.Context, params *s3.HeadObjectInput, optFns ...func(*s3.Options)) (*s3.HeadObjectOutput, error)
+	ListObjects(ctx context.Context, params *s3.ListObjectsInput, optFns ...func(*s3.Options)) (*s3.ListObjectsOutput, error)
+	GetObject(ctx context.Context, params *s3.GetObjectInput, optFns ...func(*s3.Options)) (*s3.GetObjectOutput, error)
+	CreateBucket(ctx context.Context, params *s3.CreateBucketInput, optFns ...func(*s3.Options)) (*s3.CreateBucketOutput, error)
+	DeleteObject(ctx context.Context, params *s3.DeleteObjectInput, optFns ...func(*s3.Options)) (*s3.DeleteObjectOutput, error)
+	PutObject(ctx context.Context, params *s3.PutObjectInput, optFns ...func(*s3.Options)) (*s3.PutObjectOutput, error)
+}
 
 // S3FS is a S3 filesystem implementation.
 //
@@ -27,15 +39,17 @@ var errNotDir = errors.New("not a dir")
 // by using prefixes and delims ("/"). Because directories are simulated, ModTime
 // is always a default Time value (IsZero returns true).
 type S3FS struct {
-	cl     s3iface.S3API
-	bucket string
+	context context.Context
+	s3Api   S3Api
+	bucket  string
 }
 
 // New returns a new filesystem that works on the specified bucket.
-func New(cl s3iface.S3API, bucket string) *S3FS {
+func New(context context.Context, s3Api S3Api, bucket string) *S3FS {
 	return &S3FS{
-		cl:     cl,
-		bucket: bucket,
+		context: context,
+		s3Api:   s3Api,
+		bucket:  bucket,
 	}
 }
 
@@ -50,17 +64,17 @@ func (f *S3FS) Open(name string) (fs.File, error) {
 	}
 
 	if name == "." {
-		return openDir(f.cl, f.bucket, name)
+		return openDir(f.context, f.s3Api, f.bucket, name)
 	}
 
-	out, err := f.cl.GetObject(&s3.GetObjectInput{
+	out, err := f.s3Api.GetObject(f.context, &s3.GetObjectInput{
 		Key:    &name,
 		Bucket: &f.bucket,
 	})
 
 	if err != nil {
 		if isNotFoundErr(err) {
-			switch d, err := openDir(f.cl, f.bucket, name); {
+			switch d, err := openDir(f.context, f.s3Api, f.bucket, name); {
 			case err == nil:
 				return d, nil
 			case !isNotFoundErr(err) && !errors.Is(err, errNotDir) && !errors.Is(err, fs.ErrNotExist):
@@ -82,17 +96,17 @@ func (f *S3FS) Open(name string) (fs.File, error) {
 	}
 
 	statFunc := func() (fs.FileInfo, error) {
-		return stat(f.cl, f.bucket, name)
+		return stat(f.context, f.s3Api, f.bucket, name)
 	}
 
-	if out.ContentLength != nil && out.LastModified != nil {
+	if out.LastModified != nil {
 		// if we got all the information from GetObjectOutput
 		// then we can cache fileinfo instead of making
 		// another call in case Stat is called.
 		statFunc = func() (fs.FileInfo, error) {
 			return &fileInfo{
 				name:    path.Base(name),
-				size:    *out.ContentLength,
+				size:    out.ContentLength,
 				modTime: *out.LastModified,
 			}, nil
 		}
@@ -106,7 +120,7 @@ func (f *S3FS) Open(name string) (fs.File, error) {
 
 // Stat implements fs.StatFS.
 func (f *S3FS) Stat(name string) (fs.FileInfo, error) {
-	fi, err := stat(f.cl, f.bucket, name)
+	fi, err := stat(f.context, f.s3Api, f.bucket, name)
 	if err != nil {
 		return nil, &fs.PathError{
 			Op:   "stat",
@@ -119,7 +133,7 @@ func (f *S3FS) Stat(name string) (fs.FileInfo, error) {
 
 // ReadDir implements fs.ReadDirFS.
 func (f *S3FS) ReadDir(name string) ([]fs.DirEntry, error) {
-	d, err := openDir(f.cl, f.bucket, name)
+	d, err := openDir(f.context, f.s3Api, f.bucket, name)
 	if err != nil {
 		return nil, &fs.PathError{
 			Op:   "readdir",
@@ -130,23 +144,23 @@ func (f *S3FS) ReadDir(name string) ([]fs.DirEntry, error) {
 	return d.ReadDir(-1)
 }
 
-func stat(s3cl s3iface.S3API, bucket, name string) (fs.FileInfo, error) {
+func stat(context context.Context, s3Api S3Api, bucket, name string) (fs.FileInfo, error) {
 	if !fs.ValidPath(name) {
 		return nil, fs.ErrInvalid
 	}
 
 	if name == "." {
 		return &dir{
-			s3cl:   s3cl,
-			bucket: bucket,
+			context: context,
+			s3Api:   s3Api,
+			bucket:  bucket,
 			fileInfo: fileInfo{
 				name: ".",
 				mode: fs.ModeDir,
 			},
 		}, nil
 	}
-
-	head, err := s3cl.HeadObject(&s3.HeadObjectInput{
+	head, err := s3Api.HeadObject(context, &s3.HeadObjectInput{
 		Bucket: &bucket,
 		Key:    aws.String(name),
 	})
@@ -157,25 +171,27 @@ func stat(s3cl s3iface.S3API, bucket, name string) (fs.FileInfo, error) {
 	} else {
 		return &fileInfo{
 			name:    name,
-			size:    derefInt64(head.ContentLength),
+			size:    head.ContentLength,
 			mode:    0,
 			modTime: derefTime(head.LastModified),
 		}, nil
 	}
 
-	out, err := s3cl.ListObjectsV2(&s3.ListObjectsV2Input{
+	out, err := s3Api.ListObjects(context, &s3.ListObjectsInput{
 		Bucket:    &bucket,
-		Delimiter: aws.String("/"),
-		Prefix:    aws.String(name + "/"),
-		MaxKeys:   aws.Int64(1),
+		Delimiter: aws.String(name + "/"),
+		Prefix:    aws.String(name),
+		MaxKeys:   1,
 	})
 	if err != nil {
 		return nil, err
 	}
+
 	if len(out.CommonPrefixes) > 0 || len(out.Contents) > 0 {
 		return &dir{
-			s3cl:   s3cl,
-			bucket: bucket,
+			context: context,
+			s3Api:   s3Api,
+			bucket:  bucket,
 			fileInfo: fileInfo{
 				name: name,
 				mode: fs.ModeDir,
@@ -185,8 +201,8 @@ func stat(s3cl s3iface.S3API, bucket, name string) (fs.FileInfo, error) {
 	return nil, fs.ErrNotExist
 }
 
-func openDir(s3cl s3iface.S3API, bucket, name string) (fs.ReadDirFile, error) {
-	fi, err := stat(s3cl, bucket, name)
+func openDir(context context.Context, s3Api S3Api, bucket, name string) (fs.ReadDirFile, error) {
+	fi, err := stat(context, s3Api, bucket, name)
 	if err != nil {
 		return nil, err
 	}
@@ -197,15 +213,7 @@ func openDir(s3cl s3iface.S3API, bucket, name string) (fs.ReadDirFile, error) {
 	return nil, errNotDir
 }
 
-var notFoundCodes = map[string]struct{}{
-	s3.ErrCodeNoSuchKey: {},
-	"NotFound":          {}, // localstack
-}
-
 func isNotFoundErr(err error) bool {
-	if aerr, ok := err.(awserr.Error); ok {
-		_, ok := notFoundCodes[aerr.Code()]
-		return ok
-	}
-	return false
+	var re *smithyhttp.ResponseError
+	return errors.As(err, &re) && re.HTTPStatusCode() == 404
 }
