+++ go.mod
@@ -2,4 +2,10 @@ module github.com/jszwec/s3fs
 
 go 1.16
 
-require github.com/aws/aws-sdk-go v1.36.24
+require (
+	github.com/aws/aws-sdk-go-v2 v1.13.0
+	github.com/aws/aws-sdk-go-v2/config v1.13.1
+	github.com/aws/aws-sdk-go-v2/credentials v1.8.0
+	github.com/aws/aws-sdk-go-v2/service/s3 v1.24.1
+	github.com/aws/smithy-go v1.10.0
+)
